requirements.txt
streamlit>=1.32.0
mistralai>=1.0.0
requests>=2.31.0
python-dotenv>=1.0.0
playwright>=1.40.0
